+++ wdio/step-definitions/add-networks.steps.js
@@ -1,30 +1,11 @@
 /* global driver */
-import { Given, When, Then } from '@wdio/cucumber-framework';
-import ImportFromSeedScreen from '../screen-objects/Onboarding/ImportFromSeedScreen.js';
-import MetaMetricsScreen from '../screen-objects/Onboarding/MetaMetricsScreen.js';
-import OnboardingScreen from '../screen-objects/Onboarding/OnboardingScreen.js';
-import WelcomeScreen from '../screen-objects/Onboarding/OnboardingCarousel.js';
-import Accounts from '../helpers/Accounts';
+import { When, Then } from '@wdio/cucumber-framework';
 import WalletMainScreen from '../screen-objects/WalletMainScreen';
 import AddNetworksModal from '../screen-objects/Modals/AddNetworksModal';
 import NetworksScreen from '../screen-objects/NetworksScreen';
 import NetworkApprovalModal from '../screen-objects/Modals/NetworkApprovalModal';
 import NetworkEducationModal from '../screen-objects/Modals/NetworkEducationModal';
 import NetworkListModal from '../screen-objects/Modals/NetworkListModal';
-
-Given(/^I import wallet using seed phrase "([^"]*)?"/, async (phrase) => {
-  const setTimeout = 50000;
-  await driver.pause(setTimeout);
-  await WelcomeScreen.clickGetStartedButton();
-  await OnboardingScreen.clickImportWalletButton();
-  await MetaMetricsScreen.swipeUp();
-  await MetaMetricsScreen.tapIAgreeButton();
-  const validAccount = Accounts.getValidAccount();
-  await ImportFromSeedScreen.typeSecretRecoveryPhrase(phrase);
-  await ImportFromSeedScreen.typeNewPassword(validAccount.password);
-  await ImportFromSeedScreen.typeConfirmPassword(validAccount.password);
-  await ImportFromSeedScreen.clickImportButton();
-});
 
 When(/^I tap on the Add a Network button/, async () => {
   await AddNetworksModal.clickAddNetworks();
